+++ rust/gui-client/src-common/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 
 [dependencies]
 anyhow = { version = "1.0" }
+arboard = { version = "3.4.0", default-features = false }
 atomicwrites = "0.4.3"
 connlib-shared = { workspace = true }
 crash-handler = "0.6.2"
@@ -15,10 +16,12 @@ firezone-logging = { workspace = true }
 futures = { version = "0.3", default-features = false }
 hex = "0.4.3"
 minidumper = "0.8.3"
+native-dialog = "0.7.0"
 output_vt100 = "0.1"
 png = "0.17.13" # `png` is mostly free since we already need it for Tauri
 rand = "0.8.5"
 reqwest = { version = "0.12.5", default-features = false, features = ["stream", "rustls-tls"] }
+sadness-generator = "0.5.0"
 secrecy = { workspace = true }
 semver = { version = "1.0.22", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
