+++ rust/gui-client/src-tauri/src/client/elevation.rs
@@ -2,8 +2,8 @@ pub(crate) use platform::gui_check;
 
 #[cfg(target_os = "linux")]
 mod platform {
-    use crate::client::gui::Error;
     use anyhow::{Context as _, Result};
+    use firezone_gui_client_common::errors::Error;
     use firezone_headless_client::FIREZONE_GROUP;
 
     /// Returns true if all permissions are correct for the GUI to run
@@ -36,8 +36,8 @@ mod platform {
 
 #[cfg(target_os = "windows")]
 mod platform {
-    use crate::client::gui::Error;
     use anyhow::Result;
+    use firezone_gui_client_common::errors::Error;
 
     // Returns true on Windows
     ///
